+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+3.4.0
+-----
+
+ * deprecated the ability to check for the initialization of a private service with the `Container::initialized()` method
+
 3.3.0
 -----